move texture caching jobs to a separate file
